Initial progress with the webapp
Added significant functionality to the webapp. It now accepts a user's quiz title, and display the quiz builder page. The user can create and store questions and finish the quiz, which returns to the main screen. Functionality that needs to be added, is the ability to add a second quiz without impacting the first, the ability to play the quiz and the ability to edit the quiz.